+++ src/extension.ts
@@ -164,9 +164,9 @@ export function activate(context: vscode.ExtensionContext): void {
     { scheme: 'file', pattern: '**/Directory.{Packages,Build}.props' },
     { scheme: 'file', pattern: '**/packages.config' },
     { scheme: 'file', pattern: '**/Gemfile' },
-    { scheme: 'file', pattern: '**/*.gemspec' },
     { scheme: 'file', pattern: '**/mix.exs' },
     { scheme: 'file', pattern: '**/*.tf' },
+    { scheme: 'file', pattern: '**/*.tofu' },
   ];
 
   context.subscriptions.push(